# fam_tree_rest_api/api/src/models/individuals.rs
use chrono::NaiveDate;
use serde::{Deserialize, Serialize};
use sqlx::FromRow;

use super::{Id, Individual};

#[derive(Serialize, Deserialize, Clone, FromRow, Debug)]
pub struct Individuals {
    pub id: Id,
    pub first_name: String,
    pub last_name: String,
    pub gender: String,

    #[serde(with = "date_format")]
    pub date_of_birth: NaiveDate,
}
#[derive(Serialize, Deserialize, Clone, FromRow)]
pub struct IndividualBaseInfo {
    pub id: Id,
    pub first_name: String,
    pub last_name: String,

    #[serde(with = "date_format")]
    pub date_of_birth: NaiveDate,
}

mod date_format {
    use chrono::NaiveDate;
    use serde::{self, Deserialize, Deserializer, Serializer};

    const FORMAT: &str = "%Y-%m-%d";

    pub fn serialize<S>(date: &NaiveDate, serializer: S) -> Result<S::Ok, S::Error>
    where
        S: Serializer,
    {
        let s = format!("{}", date.format(FORMAT));
        serializer.serialize_str(&s)
    }

    pub fn deserialize<'de, D>(deserializer: D) -> Result<NaiveDate, D::Error>
    where
        D: Deserializer<'de>,
    {
        let s = String::deserialize(deserializer)?;
        NaiveDate::parse_from_str(&s, FORMAT).map_err(serde::de::Error::custom)
    }
}